Improve direct ssh invocation (#73)
Marking the plugin as not requiring a connection ensures that Ansible
won't start one.

Pulling the ssh connection parameters from the existing connection
instance has multiple advantages. It re-uses the existing and well-known
ssh variables from Ansible, which can be overridden on a task-basis if
needed. It will also use the same fallback- and merge-/priority-
strategy for the variables, and already resolve any template expressions.
The obvious downside is that it only works if the target host was
defined to be connected to via the ssh connection plugin, but in the
most cases this will be true; and if needed can be worked around using
task-level definitions in the playbook. Given the opportunity, also
incoporate other standard Ansible ssh variables into the command, to
better facilitate ad-hoc configurations and reduce the need to define
these in the local ssh config.

Defining the `conn_obj` variable prevents a potential `Variable not
defined` error in the except block, namely when pexpect fails to run
the given command.